feat: split keyboard support into default navigation + optional docking module
Keyboard accessibility was one module (AccessibilityService) covering both
everyday navigation/focus and advanced layout control. Split it so the default
navigation stays in the batteries-included bundle while the advanced docking
piece becomes an optional, separately-registered module.

dockview-modules:
- AccessibilityModule (default, in the `Modules` bundle): tab switching, F6
  group cycling, focus-the-tab-strip, and focus management (Esc-from-float,
  float Tab-trap, focus-restore-on-close). No `dependsOn`.
- KeyboardDockingModule (new, exported but NOT in the `Modules` bundle): spatial
  group focus (Ctrl+Shift+Arrows) and the Ctrl+M keyboard move/dock state machine
  with live preview + narration. `dependsOn` AdvancedDnD + LiveRegion.
- Shared keymap/opt-in helpers in keyboardShared.ts. The two document keydown
  listeners coordinate via a neutral `data-dv-kbd-moving` root attribute: while a
  move is armed the navigation listener stands down (new precedence test guards
  this).

dockview-core:
- `DockviewKeybindings` now covers the default-navigation bindings only; the
  docking/spatial bindings live in the docking module's own keymap.
- Add `IKeyboardDockingService` contract + `keyboardDockingService` service slot.

docs:
- accessibility.mdx, whats-new-v7.mdx, migrating-to-v7.mdx updated to describe
  the default navigation surface; API reference + exports snapshot regenerated.

Full suite green (1345 tests); core typechecks; docs build clean.
