+++ packages/docs/docs/overview/migrating-to-v7.mdx
@@ -92,7 +92,7 @@ set of [feature modules](/docs/core/modules) is no longer part of bare
 - **advanced drag-and-drop** (`onWillDragPanel` / `onWillDragGroup` /
   `onWillDrop` hooks, custom drag ghost / [drop overlay](/docs/core/dnd/overlay))
 - **[keyboard accessibility](/docs/advanced/accessibility)** (keyboard
-  navigation / docking / screen-reader announcements)
+  navigation / focus management / screen-reader announcements)
 
 These now live in the `dockview` package, which registers them automatically.
 **`dockview-core` is an internal package — use `dockview`** (or a framework
