+++ packages/docs/docs/overview/whats-new-v7.mdx
@@ -31,12 +31,13 @@ package is batteries-included (it registers them all), while bare `dockview-core
 ships without them. Authors can register modules explicitly with
 `registerModules` / `defineModule`.
 
-### Accessibility pack
+### Keyboard navigation & announcements
 
 A new [`keyboardNavigation`](/docs/advanced/accessibility#keyboard-navigation)
-option enables a fully rebindable keymap: tab switching, F6 / spatial group
-focus, and `Ctrl+M` keyboard docking. Layout changes can also be narrated to
-screen readers via a built-in live region, with hooks
+option enables a rebindable keymap for moving around the dock — tab switching,
+F6 group cycling, and jumping to the tab strip — alongside focus management
+(focus restore on close, floating-group focus containment). Layout changes can
+also be narrated to screen readers via a built-in live region, with hooks
 ([`announcements`](/docs/advanced/accessibility#screen-reader-announcements),
 `getAnnouncement`, `announcer`) and an i18n `messages` catalog.
